Fix level 3 errors

+++ wcfsetup/install/files/lib/system/importer/TrophyImporter.class.php
@@ -131,7 +131,7 @@ class TrophyImporter extends AbstractImporter
                         VALUES      (?, ?, ?, ?, ?)";
                 $statement = WCF::getDB()->prepare($sql);
                 $statement->execute([$objectTypeID, 0, 'Import', 0, TIME_NOW]);
-                $this->importCategoryID = WCF::getDB()->getInsertID("wcf1_category", 'categoryID');
+                $this->importCategoryID = (int)WCF::getDB()->getInsertID("wcf1_category", 'categoryID');
             }
         }
